Clean quarks link tests
