+++ docs/examples/quarks-link/native-to-bosh/README.md
@@ -1,0 +1,8 @@
+## Use Cases
+
+- [Use Cases](#use-cases)
+  - [boshdeployment.yaml](#boshdeployment)
+
+### boshdeployment
+
+This is a `BOSHDeployment` which consumes a link of a native Kubernetes `Pod` from a `Deployment`. Before creating the `BOSHDeployment` in the cluster, create the native Kubernetes requirements by creating link-secret, link-pod and link-service resources.
