+++ docs/examples/quarks-link/native-to-bosh/link-pod.yaml
