+++ docs/examples/quarks-link/native-to-bosh/link-secret.yaml
